refactor: atoms、moleculesをcomponentsに整理(#26)
