+++ src/components/molecules/CorrectAnswerCounter.tsx
@@ -1,5 +1,5 @@
 import { VFC } from 'react';
-import { Box } from '@material-ui/core';
+import Box from '@material-ui/core/Box';
 import { PanoramaFishEye as PanoramaFishEyeIcon } from '@material-ui/icons';
 
 type Props = {
